requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
distro==1.9.0
et_xmlfile==2.0.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
openai==1.59.3
openpyxl==3.1.5
packaging==24.2
paramiko==3.5.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
polars==1.19.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyNaCl==1.5.0
pypdfium2==4.30.1
pytesseract==0.3.13
python-dotenv==1.0.1
setuptools==75.6.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
wheel==0.45.1
xlsx2csv==0.8.4
XlsxWriter==3.2.0
yarl==1.18.3
